feat: implement base tileshare route; add subnav to timeline
