+++ src/core/constants/app_layout_config.ts
@@ -1,0 +1,7 @@
+const appLayoutConfig = {
+	NAV_HEIGHT: 64,
+	SUBNAV_HEIGHT: 56,
+};
+
+export default appLayoutConfig;
+export type AppLayoutConfig = typeof appLayoutConfig;
